moved test to new utils file
